+++ src/main/java/com/github/elic0de/thejpspit/listener/EventListener.java
@@ -3,7 +3,9 @@ package com.github.elic0de.thejpspit.listener;
 import com.github.elic0de.thejpspit.TheJpsPit;
 import com.github.elic0de.thejpspit.player.PitPlayer;
 import com.github.elic0de.thejpspit.player.PitPlayerManager;
+import fr.mrmicky.fastboard.FastBoard;
 import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -23,10 +25,13 @@ public class EventListener implements Listener {
 
     @EventHandler
     public void onJoin(PlayerJoinEvent event) {
-        plugin.getDatabase().ensureUser(PitPlayer.adapt(event.getPlayer())).thenRun(() ->
+        if (PitPlayerManager.isContain(event.getPlayer())) {
+            plugin.getGame().join(PitPlayerManager.getPitPlayer(event.getPlayer()));
+            return;
+        }
+        plugin.getDatabase().ensureUser(event.getPlayer()).thenRun(() ->
                 plugin.getDatabase().getPitPlayer(event.getPlayer()).join().ifPresent(pitPlayer -> {
-                    if (!PitPlayerManager.isContain(event.getPlayer())) PitPlayerManager.registerUser(pitPlayer);
-                    pitPlayer.showHealth(pitPlayer);
+                    PitPlayerManager.registerUser(pitPlayer);
                     plugin.getGame().join(pitPlayer);
                 })
         );
@@ -36,8 +41,9 @@ public class EventListener implements Listener {
     public void onQuit(PlayerQuitEvent event) {
         final PitPlayer player = PitPlayerManager.getPitPlayer(event.getPlayer());
 
+        PitPlayerManager.unregisterUser(player);
         plugin.getGame().leave(player);
-        player.getBoard().delete();
+        if (player.getBoard() != null) player.getBoard().delete();
     }
 
     @EventHandler
@@ -58,6 +64,7 @@ public class EventListener implements Listener {
 
     @EventHandler
     public void onDeath(PlayerDeathEvent event) {
+        if (event.getEntity().getKiller() == null) return;
         plugin.getGame().death(PitPlayerManager.getPitPlayer(event.getEntity()));
         event.getDrops().clear();
     }
@@ -68,17 +75,28 @@ public class EventListener implements Listener {
     }
 
     @EventHandler
-    public void onDamage(EntityDamageByEntityEvent event) {
+    public void onFallDamage(EntityDamageEvent event) {
         if (event.getEntity() instanceof Player) {
             if (event.getCause() == EntityDamageEvent.DamageCause.FALL) {
                 event.setCancelled(true);
                 return;
             }
         }
-        if (event.getDamager() instanceof Player player) {
-            final PitPlayer pitPlayer = PitPlayerManager.getPitPlayer(player);
-            pitPlayer.showHealth(pitPlayer);
+    }
+
+    @EventHandler
+    public void onDamage(EntityDamageByEntityEvent event) {
+        if (event.getEntity() instanceof Player vitim) {
+            if (event.getDamager() instanceof Player) {
+                final PitPlayer pitPlayer = PitPlayerManager.getPitPlayer(vitim);
+                if (event.getCause() == EntityDamageEvent.DamageCause.FALL) {
+                    event.setCancelled(true);
+                    return;
+                }
+                pitPlayer.showHealth(pitPlayer);
+            }
         }
+
     }
 
     @EventHandler
